Move slotable_default.rb patch to concerns folder
Also uses an updated solution from ViewComponent (the issue
has been resolved but will only be released in v4), also adds
a guard clause to remove the patch once we have upgraded
ViewComponent to `>= 4`
